Rework test execution logic for better reliability
